backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.121.0
h11==0.16.0
idna==3.11
joblib==1.5.2
numpy==2.3.4
pydantic==2.12.4
pydantic-core==2.41.5
scikit-learn==1.7.2
scipy==1.16.3
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.49.3
threadpoolctl==3.6.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0

# Novas dependências para Supabase/PostgreSQL (Async)
asyncpg==0.30.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
email-validator==2.2.0
greenlet==3.1.1
